Allows build system implementations to provide additional layer metadata
Consumers of libbs may want to augment the expected metadata with values
specific to the given implementation. For example, the maven buildpack
should invalidate the layer if the contents of settings.xml changes.

The PR also exposes ExpectedMetadata as part of the public interface of
Application to enable testing and to provide this data to consumers that
may require it.

Resolves #2

Signed-off-by: Emily Casey <ecasey@pivotal.io>